refactor(spec): extract color temporal censor recovery helpers
Move colorTemporalCensorRecovery and censoredTemporalColorRecovers into
validate-data-checks-color-temporal.ts so the color data-check orchestrator
keeps manual domain/range and sequential rules without the recovery lattice.
